requirements.txt
fastapi==0.110.2
httpx==0.27.0
pytest==8.2.0
uvicorn==0.29.0
translators==5.9.1
langchain==0.1.16
python-dotenv==1.0.1